requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Emotion modeling and analysis
scikit-learn>=1.0.0
networkx>=2.6.0

# Visualization for emotional states
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Natural language processing for protocol analysis
nltk>=3.6.0
spacy>=3.2.0
transformers>=4.20.0

# Configuration and data handling
pyyaml>=5.4.0
jsonschema>=3.2.0

# API and interface
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Testing and validation
pytest>=6.2.0
pytest-cov>=2.12.0
hypothesis>=6.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
click>=8.0.0
rich>=10.0.0
python-dotenv>=0.19.0